requirements.txt
streamlit
opencv-python
numpy
tensorflow
requests
time
json
paho.mqtt